+++ docs/versionhistory.rst
@@ -11,6 +11,8 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
   asyncio and curio
 - Fixed task cancellation not happening right away on asyncio and curio when a cancel scope is
   entered when the deadline has already passed
+- Fixed exception group containing only cancellation exceptions not being swallowed by a timed out
+  cancel scope on asyncio and curio
 - Added the ``current_time()`` function
 
 **1.0.0rc1**